requirements.txt
beautifulsoup4==4.11.2
dnspython==2.3.0
docx==0.2.4
Flask==2.2.2
Flask_Login==0.6.2
opencv_python==4.5.5.64
openpyxl==3.0.9
pandas==1.3.5
PyPDF2==3.0.1
python_docx==0.8.10
python_magic==0.4.27
requests==2.28.2
